Refactor imports and improve component structure in ChallengesList and ImageUpload
